Convert homepage back to server rendering
