+++ pages/columns/utils.ts
@@ -1,16 +1,4 @@
-import { apiV2Prefix } from "@macrostrat-web/settings";
-import fetch from "cross-fetch";
-
-export async function fetchAPIData(apiURL: string, params: any) {
-  let url = new URL(apiV2Prefix + apiURL);
-  if (params != null) {
-    url.search = new URLSearchParams(params).toString();
-  }
-  console.log(url.toString());
-  const res = await fetch(url.toString());
-  const res1 = await res.json();
-  return res1?.success?.data;
-}
+import { fetchAPIData } from "~/_utils/fetch-helpers";
 
 export type ColumnGroup = {
   id: number;
