+++ pages/projects/+data.ts
@@ -1,4 +1,4 @@
-import { fetchAPIData } from "../columns/utils";
+import { fetchAPIData } from "~/_utils";
 
 export async function data(pageContext) {
   // `.page.server.js` files always run in Node.js; we could use SQL/ORM queries here.
